rename registrar provider to Azure, and extract migrating provider class
