+++ registration/app/registration/models/LegacyRegistration.scala
@@ -3,6 +3,7 @@ package registration.models
 import play.api.libs.json._
 import play.api.libs.json.JodaReads._
 import LegacyJodaFormat._
+import models.Provider
 
 object LegacyJodaFormat {
   val DateTimePattern = "yyyy-MM-dd'T'HH:mm:ss'Z'"
@@ -54,7 +55,8 @@ case class LegacyPreferences(
   edition: String,
   teams: Option[Seq[String]],
   matches: Option[Seq[LegacyMatch]],
-  topics: Option[Seq[LegacyTopic]]
+  topics: Option[Seq[LegacyTopic]],
+  provider: Option[Provider]
 ) {
   def hasNewsstand: Boolean = topics.exists(_.contains(LegacyTopic.NewsstandIos))
   def withNewsstand: LegacyPreferences = {
